fix(server): restore v1 notes.py, move baoku notes to baoku_notes.py
The previous commit overwrote the existing v1 notes.py (upstream proxy
adapter), breaking 13 tests. Restored the original and moved the new
local-DB notes CRUD to baoku_notes.py, registered only in the baoku
server.
